+++ docs/for-developers/CHANGELOG.md
@@ -7,8 +7,7 @@ nav_order: 4
 
 ## [Unreleased]
 
-- Change nick on server login?
-- Add colorette and other lib/_setup stuff.
+- Add in @file header type documentation.
 - Add in all the Sapphire plugins, cos why not.
 - Will the theme scrollbar the nav when 100 games exist?
 - Don't run a game if private game data has not been filled in?
@@ -19,4 +18,5 @@ nav_order: 4
 
 ### Added
 
+- Prisma support stubbed out.
 - GitHub Pages documentation stubbed out.